Disable gzip compression of application/json data in nginx.conf.
Gzip compression causes the Etag header to be stripped, as the Etag has to
exactly match the HTTP response body -- this currently breaks the Garden
dashboard, which expects to be able to HEAD a document and fetch its Etag.

Long-term solution is probably to have Garden send an `Accept:` header that
prohibits compression. Until then, we've disabled compression for data of type
application/json.