Copy the compiler's dynamic libssp.so to initrd.
Some packages (namely, `avahi`) force dynamic linkage to GCC's `libssp` even if
your local settings are requesting static linkage to `libssp.a`. Go ahead and
include the dynamic library in the initrd; after this change lands in master, we
could probably convert the whole build over to a dynamic `libssp` just to save a
bit of space.