Revert server-api nginx.conf changes for this branch.
We're going to keep these in `dev` for the time being; the next MMVA release
will include the server-api support (after we solve the issue of not being able
to install apps via the dashboard, i.e. add a special rule so that everything
unser `/dashboard` goes directly to the `couchdb` proxy target).