Always store `*.pid` files on volatile storage.
This updates the service supervision scripts, `nginx`, and `avahi` to keep its
PID and run control / state files on temporary in-memory storage that is
automatically erased on reboot. These files have no meaning outside of a
running system.

We're doing this as a defensive measure. An unclean shutdown can occasionally
leave a PID file truncated but not unlinked; Avahi, in this case, has a bug
where it interprets the empty file to mean PID zero (likely because of sloppy
`atoi` usage or missing error checking for `strtoul`) and thinks there's another
copy already running. For historical reasons, PID zero has special meaning in
UNIX-like operating systems (it's sometimes referred to as the scheduler or the
"swapper", and is actually part of the kernel rather than a normal process).

More information on the Avahi bug:
  https://bugzilla.redhat.com/show_bug.cgi?id=1050201

I'm hopeful that this will resolve issue #5:
  https://github.com/medic/medic-os/issues/5