Unset CPPFLAGS when building gcc, as it breaks libiberty.
It seems that having `CPPFLAGS` set to `-D_FORTIFY_SOURCE=2` causes the
libiberty `configure` script to think `limits.h` is unavailable when building
with gcc v4.8.2. It likely has to do with a warning being mistaken for an error
in the script (namely, `gcc` warns that optimization must be turned on).