Add `/setup/poll` API.
This API endpoint allows the requestor to determine if all of the
required background services (currently just `medic-api`) have started
up successfully. If everything required is already running, then an
object of the form `{ "ready": true, "detail": "..." }` is returned.
If one or more required background services are not running (or have not
yet finished starting up), the `ready` property will be false.

There's a `detail` property in every response; this is a string that
contains a human-readable message.

If `ready` is `true`, the response may contain additional properties
describing the background services; these properties are not final and
are subject to change (e.g. `version`).

The concierge application itself isn't set up to use this yet; that
(plus a user interface refresh) will be part of a subsequent commit.