Check in concierge's `node_modules` directory.
I'm hesitant to do this, and I know it's sometimes seen as poor form,
but we've done a lot of work to trim these dependencies down to a
minimal size (12MiB vs. 44MiB originally). Checking the directory in
gives us the ability to do things like manual deduplication, or removal
of build-time test data. It'd be nicer if we could script those
transformations, but many of these modules are moving very quickly,
and I'm not sure we could keep up.