Always enable installation on first run.
Gardener may be starting up for the first time. Always enable
installation in the first queued poll operation, so that modules
attached to the design document but not yet seen by gardener can be
installed and started properly.

For every other module that's managed by dashboard, at every other time:
always defer to dashboard for the "okay to start" signal, which is
delivered via the database changes feed (using `follow`).

Will merge into the gardener tree as well.