Remove 15MiB ooxml-schemas JAR file from distribution.
This is a dependency of Tika, which is a dependency of Lucene, which is
a dependency of couchdb-lucene. A runtime trace shows that neither it or
Tika are used currently in normal operation, but they're used by
couchdb-lucene if someone uploads a Microsoft Office document as an
attachment for indexing.

We'll find a better way to disable the indexing of Microsoft Office
documents altogether soon. Until then, this keeps the size of the
distribution sane.