Toward uptime/restart statistics.
Modify the supervisor scripts to maintain accounting data for supervised
processes, including number of restarts and seconds of uptime.

This is a work in progress; it's mostly done, but needs more testing and some
additional cleanup when services are brought down or removed. I'll finish that
up later today, and modularize `supervisor-info` a bit.