Add Changelog, rewrite README, widen version space.
The package version numbers now employ two digits for the revision number, two
digits for the minor number, and one or more digits for the major number (in
order from least to most significant digits). The version string 3.15.3 would be
represented as 31503. We're using an integer only because it's easier for
command-line jobs and shell scripts to deal with.