Mark unprintable ranges in prompt.
The Busybox line-editor source code contains the following comment: "lineedit
does not know that the terminal escape sequences do not take up space on the
screen. The redisplay code assumes, unless told otherwise, that each character
in the prompt is a printable character that takes up one character position on
the screen.  You need to tell lineedit that some sequences of characters in the
prompt take up no screen space. Compatible with readline; use the \[ escape to
begin a sequence of non-printing characters, and the \] escape to signal the
end of such a sequence.

Use these, so that the line editor input wraps in the right place.