Delete unknown users: mostly finished.
This makes the user creation process repeatable, by seeking out and deleting any
previously-created administrative users and/or user documents. The latter isn't
yet working, because a `rev` is required for a successful user `DELETE`
operation. The next commit should solve that problem, as well as turn on the
view regeneration code. After that, the backend code should be done, and all
that'll be left is some minor UI reworking.