Add /srv/system and move some internal files there.
This commit adds a `/srv/system` directory, which is further subdivided along
the usual lines (e.g. `storage`, `settings`). This directory stores internal
data that's not intended to be modified by users or system administrators.
The plan is to use `/srv/system/settings/packages` to store rewritten
configuration files, and leave the user-modifiable configuration files (now
optionally including template directives) in `/srv/settings`.

This commit also removes many of the hard-coded instances of `/srv`, in hopes of
eventually storing this in a configuration file somewhere, and allowing the
scripts to work in any base directory.