Prohibit multiple template file extensions.
If we encounter a file with a name like `local.ini.t.t`, just ignore it.
Interpreting it would produce `local.ini.t`; that output file could overwrite
another template, or we'd race with `find` and possibly reinterpret that output
file (to produce `local.ini`).

It's tempting to raise an error, allowing the condition to be quickly identified
and hopefully fixed... but doing so could easily cause a boot to fail that would
have otherwise worked in spite of the file. Best thing to do here is ignore the
file and move on (and perhaps add a warning in the future).