Package-specific configuration `->` envdir format.
Remove the weird package-specific `configuration` directory, and replace it with
an `envdir`-style directory of environment variable name/value pairs called
`env`. This directory will sit in `/srv/settings/$package/env`, and be used for
any post-deployment configuration of package-specific options. We currently just
use it to set `NODE_ENV`, but it'll eventually be useful for feeding options in
to configuration-file templates.