Forcibly disable libcap for Avahi.
If `libcap` is present on the build system, Avahi will link `avahi-daemon`
against it. We don't (yet) ship libcap with the initrd, and avahi has no
built-in "friendly" way to disable it. Set `ac_cv_header_sys_capability_h`
to `no`, which will trick `configure` in to thinking the `capability.h` kernel
header isn't present, which in turn will disable linking against `libcap`.

If we ever start shipping `libcap` with the initrd, we can back this patch out.