Revert to usbutils-007 in master.
The usbutils-008 package now requires libudev.so.1 version 196+, which is
inconveniently now part of systemd. We're currently still using Tinycore
binaries for udevd and udevadm, which are slightly older and don't ship with a
libudev shared library.

To do: replace these with Gentoo's standalone eudev package in the `develop`
branch, then update to usbutils-008 and jump through a few hoops to get the
linkage right regardless of the status of udev on the build system.