Use our own compiler even if `cc` is invoked.
Also, remove the useless CFLAGS definition beneath `source` and move
it in to the `core` package where it's actually needed. Use the
`-fstack-protector-all` option within `core` when possible, since
the base system binaries (like `dhcpcd`) tend to have some level
of network exposure and are more likely to be privileged.