Use `pipefail` in all helper scripts.
This is basically a best practice; the default is for a pipeline to exit
with the status of its final command, which is almost certainly not what
anybody wants if they're trying to handle errors sanely. The `pipefail`
option causes a pipeline to exit with the leftmost non-zero status.

This fixes a few issues in the auto-download process, especially if the
system is missing required tools.