Explicitly require root.
It'd take quite a bit of effort to compile as an unprivileged user but
install and make system modifications as root. It's also rather low
priority, since we're almost always running these builds on disposable
VMs or instances. Explicitly require root, at least for now, to avoid
ugly error messages in the case of a mistake.

There's probably a way to selectively drop privileges during
`./configure` and `make` in the `builder` scripts, but we can deal with
that when/if it's actually needed.