Always use #!/bin/bash at build time.
The build-time scripts use bash-isms, most of which are supported by shells like
ash, with the exception of $PIPESTATUS. Of course, God only knows what the
default shell is (Ubuntu uses dash), so just be explicit and request bash.

This doesn't apply to the runtime scripts; all of those restrict themselves to
features available in the Busybox version of ash.