Fix Maven build.
Remove the target directory before building, so that reinstalls are safe.
Currently, the Maven build will crash if installed files are present from a
prior run.

Fix an ancient script but, wherein the exit status of `yes` was being taken
rather than the exit status of the actual `ant` bootstrap process. Turns out
that `yes` likes to exit with status 141.