Always wait before attempting a service restart.
Before this commit, we attempted to immediately restart any service that exited
with a zero status. This didn't make a lot of sense, since a malfunctioning
service might erroneously exit zero, and services aren't supposed to be exiting
at all. Further, it hammered the CPU if a service immediately exited zero. Wait
between restarts regardless of the exit status.

This commit also reduces the wait period from five seconds to two. We might go
lower if there's a need.