Refactor: rebuild views before gardener starts.
This commit reorganizes the setup process, in order to allow service account
creation to occur *after* CouchDB views are rebuilt.

This was originally prompted by a first-run failure: the theory was that view
regeneration would cause requests from medic-api and/or medic-sentinel to block,
stack up, and eventually overwhelm CouchDB/Erlang's client/process limit. After
some post-changeset testing, it looks like this may not capture the entire
problem - but the refactor fixes the original symptom, cleans code up
significantly, and provides better feedback during the polling phase.

To do: this changeset breaks the ability to perform an unauthenticated reset of
the administrator password. We'll probably have to fix this by splitting the
`service` account in to two accounts: one for internal use by concierge, and one
for use by all other services. This approach would allow us administrative
access to CouchDB (i) without knowing the old user-created `admin` password, and
(ii) without triggering a start of `gardener` / `medic-api` / `medic-sentinel`.
The support code required to do this is already refactored; I'll fix this to-do
before I merge the branch.