Split `service` account into two accounts.
This changeset splits the `service` account in to two different accounts:
`concierge`, which is used internally by concierge to ensure it always has
administrative access to CouchDB, and `service`, which is used by local services
to connect to CouchDB (and as a cue for those services to start).

This should fix the password reset issue mentioned in the previous commit.

Note: An administrative password reset currently deletes any administrative
users that are unknown to concierge. In an emergency password reset scenario on
a production host - say, if someone had SSH access but didn't know how to reset
the CouchDB administrative password, and so they ran `svc-enable concierge` -
this could be a very bad thing. Need to think about how best to fix that.