Split the `service` account up, fix regression.
Break the `service` account up in to `gardener`, `lucene`, and `transport`
accounts, so that we have finer-grained control of service startup timing. This
will let couchdb-lucene to start earlier, which in turn will allow us to have it
regenerate indexes at the same time as CouchDB itself.

Fix a regression where the randomly-generated service password was written to
disk before the password was actually set. If the database server experienced a
transient failure and was unable to satisfy the password-setting request, the
on-disk file would stay and cause 401s on subsequent password-setting
operations. Flip the order around; local disk writes are (probably) less
likely to fail.