Simplest-possible CLI log exporter. Resolves #905.
Run `sudo /boot/print-logs` as `gardener` to export a collection of recent
server logs as Markdown on standard output. Currently, no arguments are
accepted, and the exporter prints the most recent (not-yet-rotated) chunk of
every `*.log` file in a `logs` directory.  In the future, this will accept
arguments that (i) limit the export scope to a particular package/service, (ii)
specify the number of log file segments to export for each service; and (iii)
allow the title/banner to be suppressed.