Disable systemd "predictable device naming".
We just want the first network adapter. We don't care where it is, what
slot it's in, or what kind of crazy hardware topology it's connected to.

This fixes an issue where a VMware network adapter would appear as
`eth0` on x86_64 and `eno16777736` on x86, all else being equal.