Toward medic/medic-webapp#1016.
Add currently-unvalidated phone and e-mail fields for user testing purposes.
These values are currently connected to fields in the `user` document, but are
not used to create any new contacts and/or facilities. Other minor text and
user interface changes.

Hidden by default post-setup password readback will be part of the next
changeset.