Add GNU Nano to medic-core
The built-in Busybox `vi` has a couple of quirks that makes it
really annoying for humans to use, especially if they're not already
to vim or far-less-good-than-vim variants.

This patch should create a net cost savings, primarily arising from
a reduction in broken keyboards and/or computer equipment.

This won't be available without medic-core installed, but that's
probably okay IMO.

Signed-off-by: David Brown <david@medicmobile.org>