Fix compilation of GCC-required `isl` and `cloog`
The `cloog` package has a bundled version of `isl` that it
uses by default; this bundled version helpfully causes the ISL
version string to end up as `UNKNOWN`, which breaks GCC's
autoconf script.

Tell `cloog` to use our version of `isl`, not its bundled
version. Not entirely sure why this issue was being avoided
in the past, but it definitely was happening on OS X 10.11.

Signed-off-by: David Brown <david@medicmobile.org>