Add package manifest manipulation helpers
This is a set of fairly simple shell functions that can be used
to compare and manipulate package manifests. The approach is
painfully simplistic and low-tech, but it's capable of at least
comparing a newly-installed package manifest to the previous version,
identifying differences between the two, and filtering those
differences to limit the results to only files or only directories.

This can be used for Slackware-style two-phase package installation:
install the new package over the old one, compare manifests to remove
any previously-installed files that no longer exist in the new package
(subject to exclusions for things like configuration files), and then
restart the affected services.