Support for automatic file/directory removal
Local merge: use the package manifest to guide file and directory
removal, rather than relying on a removal script for this basic
functionality. When updating a package, pre-install it over the
old version, compare the old manifest to the new manifest, and
auto-remove any files that are no longer needed.