Bug fixes for gardener and medic-core packages
In medic-core, detect when we're updating rather than installing for
the first time. In the former case, don't blow away the entire CouchDB
user database and all application data. We're definitely going to want
to keep that stuff.

In gardener, avoid use of `nohup` for a while, as it doesn't appear
to be strictly necessary and may be causing an issue with gardener
that I'm still troubleshooting.