Use sticky bit on files to ignore during removal
This is a bit of a hack, but I'm going to try it out: use the
sticky bit on files to denote that they've been modified locally
and shouldn't be auto-removed by the package management system.
May reverse this back out later if it's not useful, is too
unwieldly, or there ends up being a better way to do it.