Temporarily enable setup script output
This changeset makes output from package setup scripts visible when
installing/updating packages from the command line.

Some future work needs to be done to ensure this sort of output is
displayed to an interactive user, hidden from a noninteractive user,
and always logged persistently to disk. We likely need to replace
the `&> "$log"` target with a pipe-based target like `| "$log_cmd" 2>&1`,
pick a location for system-level output that's not otherwise captured
by a script, and then make sure we don't leak any output to the
system console. This can be put off for a bit.