Build a medic-couch2pg package with dependencies
Add the medic-couch2pg package, with dependencies installed at build-time
using a `builder` style script (like `medic-core`).  Some minor
refactoring to make this possible. Consider unbundling dependencies for
`medic-transport` at some point in the future, using this package as a
guide.

If installed, the package likely will eternal-restart - at least until
end-of-day tomorrow, when I have a chance to get it on a live system and
make sure credentials and connectivity parameters are being set up
correctly in the startup script environment.