Lock major versions of outside `npm -g` modules
Async 2.x was released with breaking API changes; using it with
gardener against the advice of package.json breaks spectacularly.
The differences seem to be in `detectSeries` and `auto`; I spent
a bit of time putting together a compatibility patch for gardener,
but have decided to defer it for now. Use async v1.x for now.