Mitigate `npm`-associated build failure
This is a fun one: it looks like `npm` is doing something to
standard output that causes later shell-based pipeline steps
(e.g. a `while` loop) to fail when writing to stdout with
"write error: Resource temporarily unavailable". This happens
on Linux, but not on OS X.

Avoid connecting the build script's standard output to another
shell for now - connect it directly to `pv`. This gets past the
problem and lets builds continue running to completion.

Here's a minimal repro case that illustrates the problem on
Amazon Linux 2016.03. It doesn't encounter the write error
on a fully-patched OS X:

```
$ npm install npm &>/dev/null
$ echo ${PIPESTATUS[*]}
$ (cp -av node_modules new) | while read x; do a=a; done
$ echo ${PIPESTATUS[*]}
$ (npm list) | while read x; do a=a; done
$ echo ${PIPESTATUS[*]}
$ (npm list; cp -av node_modules new) | while read x; do a=a; done
cp: write error
$ echo ${PIPESTATUS[*]}
```

I'll take a closer look this weekend, and will probably learn a
lot of horrible 70s terminal trivia while I'm there.