Add top-level `compile-only` target
Toward medic/medic-webapp#3023. This provides a top-level
`make compile-only` target that does the full build, but no
size-optimization or post-build deletion steps, and fully
maintains any installed documentation. Use this to test
the documentation packaging script, and then invoke that
in the right spot.

This is useful for a whole bunch of other things, and
provides a nice VM-snapshottable mid-way point for tuning
the size optimization and all of the post-build/packaging
steps.