Add X-Forwarede-Proto: https; to all proxy rules
This fixes a replication timeout that we observed during an
upgrade from 2.8.5 to 2.11.0 on an LG training machine.

We haven't seen this issue on any other upgrade instance.
The onluy difference that comes to mind is that we were not
using `update_markets.sh` to switch channels in this case.

Root cause is still unclear; will investigate further.
In the mean time, this changeset allows dashboard to update
properly.