Add libtinfo symlinks back
These work when they're in /lib, but there are still obviously
some unanswered questions (and a library versioning warning).
Add this to the end of the to-do list for now and continue on.

This was, in fact, the underlying failure mode: the `timestamp`
script uses `bash` in order to correctly handle signals and
subprocess termination; linking errors with libtinfo.so broke
startup scripts, which in turn broke several services.