Raise default timeout
This isn't anywhere near ideal from a security perspective,
but it is what it is. Raise the default session timeout and
use local.ini to lower it if needed. This should avoid any
scenario where users are accidentally logged out due to a
change control or human error issue in configuration.