Fix all known first-run issues that are external to horticulturalist
The checklist in medic/medic-webapp#4157 is complete. We've found a
couple issues that appear to involve horticulturalist code itself, so
I've split off medic/medic-webapp#4204 to track that work. Ideally,
it'll be minor, and then I'll generate a PR for the Medic OS
landscaping work and one in `horticulturalist` for some user/perm
related issues that currently block bootstrapping in Docker.